+++ Setup/UpgradeData.php
@@ -1,0 +1,71 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+namespace Alekseon\CustomFormsBuilder\Setup;
+
+use Alekseon\AlekseonEav\Model\Adminhtml\System\Config\Source\Scopes;
+use Magento\Framework\Setup\ModuleContextInterface;
+use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Magento\Framework\Setup\UpgradeDataInterface;
+
+/**
+ * Class UpgradeData
+ * @package Alekseon\CustomFormsBuilder\Setup
+ */
+class UpgradeData implements UpgradeDataInterface
+{
+    /**
+     * @var EavDataSetupFactory
+     */
+    protected $eavSetupFactory;
+    /**
+     * @var \Alekseon\CustomFormsBuilder\Model\Form\AttributeRepository
+     */
+    protected $formAttributeRepository;
+
+    /**
+     * InstallData constructor.
+     * @param \Alekseon\AlekseonEav\Setup\EavDataSetupFactory $eavSetupFactory
+     * @param \Alekseon\CustomFormsBuilder\Model\Form\AttributeRepository $formAttributeRepository
+     */
+    public function __construct(
+        \Alekseon\AlekseonEav\Setup\EavDataSetupFactory $eavSetupFactory,
+        \Alekseon\CustomFormsBuilder\Model\Form\AttributeRepository $formAttributeRepository
+    ) {
+        $this->eavSetupFactory = $eavSetupFactory;
+        $this->formAttributeRepository = $formAttributeRepository;
+    }
+
+    /**
+     * @param ModuleDataSetupInterface $setup
+     * @param ModuleContextInterface $context
+     */
+    public function upgrade(ModuleDataSetupInterface $setup, ModuleContextInterface $context)
+    {
+        if (version_compare($context->getVersion(), '1.0.1', '<')) {
+            $this->addEnableEmailNotificationAttribute($setup);
+        }
+    }
+
+    /**
+     * @param $setup
+     */
+    protected function addEnableEmailNotificationAttribute($setup)
+    {
+        $eavSetup = $this->eavSetupFactory->create();
+        $eavSetup->setAttributeRepository($this->formAttributeRepository);
+
+        $eavSetup->createAttribute(
+            'enable_email_notification',
+            [
+                'frontend_input' => 'boolean',
+                'frontend_label' => 'Notify By Email about new entities',
+                'visible_in_grid' => true,
+                'sort_order' => 100,
+                'scope' => Scopes::SCOPE_GLOBAL,
+            ]
+        );
+    }
+}