+++ CHANGELOG.md
@@ -7,8 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 ### Added
 
-## [Unreleased]
+## [102.0.1] - 2020-04-01
 ### Added
+- new entity notifiaction email
 - option sources for select attributes
 - form record created at column
 - record attribute variables for events